% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_k.R
\name{approx_k}
\alias{approx_k}
\alias{approx_k.DocumentTermMatrix}
\alias{approx_k.TermDocumentMatrix}
\title{Approximate Number of Clusters for a Text Matrix}
\usage{
approx_k(x, verbose = TRUE)

\method{approx_k}{TermDocumentMatrix}(x, verbose = TRUE)

\method{approx_k}{DocumentTermMatrix}(x, verbose = TRUE)
}
\arguments{
\item{x}{A matrix.}

\item{verbose}{logical.  If \code{TRUE} the k determination is printed.}
}
\value{
Returns an integer.
}
\description{
Can & Ozkarahan (1990) formula for approximating the number of clusters for
a text matrix: \eqn{(m * n)/t} where \eqn{m} and \eqn{n} are the dimensions
of the matrix and \eqn{t} is the length of the non-zero elements in matrix
\eqn{A}.
}
\examples{
library(gofastr)
library(dplyr)

presidential_debates_2012 \%>\%
    with(q_dtm(dialogue)) \%>\%
    approx_k()
}
\references{
Can, F., Ozkarahan, E. A. (1990). Concepts and effectiveness of
the cover-coefficient-based clustering methodology for text databases.
ACM Transactions on Database Systems 15 (4): 483. doi:10.1145/99935.99938. \cr
}

