% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cluster.R
\name{assign_cluster}
\alias{assign_cluster}
\alias{assign_cluster.hierarchical_cluster}
\alias{assign_cluster.kmeans_cluster}
\alias{assign_cluster.nmf_cluster}
\alias{assign_cluster.skmeans_cluster}
\title{Assign Clusters to Documents/Text Elements}
\usage{
assign_cluster(x, k = approx_k(get_dtm(x)), h = NULL, ...)

\method{assign_cluster}{hierarchical_cluster}(x, k = approx_k(get_dtm(x)),
  h = NULL, ...)

\method{assign_cluster}{kmeans_cluster}(x, ...)

\method{assign_cluster}{skmeans_cluster}(x, ...)

\method{assign_cluster}{nmf_cluster}(x, ...)
}
\arguments{
\item{x}{a \code{xxx_cluster} object.}

\item{k}{The number of clusters (can supply \code{h} instead).  Defaults to
use \code{approx_k} of the \code{\link[tm]{DocumentTermMatrix}} produced
by \code{data_storage}.}

\item{h}{The height at which to cut the dendrograms (determines number of
clusters).  If this argument is supplied \code{k} is ignored.}

\item{\ldots}{ignored.}
}
\value{
Returns an \code{assign_cluster} object; a named vector of cluster
assignments with documents as names.  The object also contains the original
\code{data_storage} object and a \code{join} function.  \code{join} is a
function (a closure) that captures information about the \code{assign_cluster}
that makes rejoining to the original data set simple.  The user simply
supplies the original data set as an argument to \code{join}
(\code{attributes(FROM_ASSIGN_CLUSTER)$join(ORIGINAL_DATA)}).
}
\description{
Assign clusters to documents/text elements.
}
\examples{
library(dplyr)

x <- with(
    presidential_debates_2012,
    data_store(dialogue, paste(person, time, sep = "_"))
)

hierarchical_cluster(x) \%>\%
    plot(h=.7, lwd=2)

hierarchical_cluster(x) \%>\%
    assign_cluster(h=.7)

hierarchical_cluster(x, method="complete") \%>\%
    plot(k=6)

hierarchical_cluster(x) \%>\%
    assign_cluster(k=6)


x2 <- presidential_debates_2012 \%>\%
    with(data_store(dialogue)) \%>\%
    hierarchical_cluster()

ca2 <- assign_cluster(x2, k = 55)
summary(ca2)

## add to original data
attributes(ca2)$join(presidential_debates_2012)

## split text into clusters
get_text(ca2)

## Kmeans Algorithm
kmeans_cluster(x, k=6) \%>\%
    assign_cluster()

x3 <- presidential_debates_2012 \%>\%
    with(data_store(dialogue)) \%>\%
    kmeans_cluster(55)

ca3 <- assign_cluster(x3)
summary(ca3)

## split text into clusters
get_text(ca3)
}

