% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustext-package.R
\docType{data}
\name{assignments}
\alias{assignments}
\title{Topic Assignments}
\format{A list with 3 elements}
\usage{
data(assignments)
}
\description{
A dataset containing a list of topic assignments by various clustering 
algorithms.  Assignments correspond to the rows (minus empty rows) of the 
\code{presidential_debates_2012} data set.
}
\keyword{datasets}

