% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Merge Clusters & Cluster Categories Back to Original Data}
\usage{
categorize(data, assign.cluster, cluster.key)
}
\arguments{
\item{data}{A data set that was fit with a cluster model.}

\item{assign.cluster}{An \code{\link[clustext]{assign_cluster}} object.}

\item{cluster.key}{An \code{\link[clustext]{assign_cluster}} object.}
}
\value{
Returns a \code{\link[base]{data.frame}} key of clusters and categories.
}
\description{
Merge clusters, categories, and the original data back together.
}
\examples{
library(dplyr)

## Assign Clusters
ca <- presidential_debates_2012 \%>\%
    with(data_store(dialogue)) \%>\%
    hierarchical_cluster() \%>\%
    assign_cluster(k = 7)

## Write Cluster Text for Human Categorization
write_cluster_text(ca)
write_cluster_text(ca, n.sample=10)
write_cluster_text(ca, lead="  -", n.sample=10)

## Read Human Coded Categories Back In
categories_file <- system.file("additional/foo_turk.txt", package = "clustext")
readLines(categories_file)
(categories_key <- read_cluster_text(categories_file))

## Add Categories Back to Original Data Set
categorize(
    data = presidential_debates_2012,
    assign.cluster = ca,
    cluster.key = categories_key
)
}
\seealso{
\code{\link[clustext]{write_cluster_text}},
\code{\link[clustext]{read_cluster_text}}
}

