% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustext-package.R
\docType{package}
\name{clustext}
\alias{clustext}
\alias{clustext-package}
\alias{package-clustext}
\title{Consistent Clustering for Text Data}
\description{
Optimized, consistent tools for clustering text data.
}

