% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Adjusted Rand Index Comaprison Between Algorithms}
\usage{
compare(...)
}
\arguments{
\item{\ldots}{A series of outputs from \code{assign_cluster} for various
cluster algorithmns.}
}
\value{
Returns a pair-wise comparison matrix of Adjusted Rand Indices for
algorithm.  Higher Adjusted Rand Index scores indicate higher cluster
assignment agreement.
}
\description{
An Adjusted Rand Index comparison of the assignments between different
clustering algorithms.
}
\examples{
compare(
    assignments$hierarchical_assignment,
    assignments$kmeans_assignment,
    assignments$skmeans_assignment,
    assignments$nmf_assignment
)

## Understanding the ARI
set.seed(10)
w <- sample(1:10, 40, TRUE)
x <- 11-w
set.seed(20)
y <- sample(1:10, 40, TRUE)
set.seed(50)
z <- sample(1:10, 40, TRUE)

data.frame(w, x, y, z)

library(mclust)

mclust::adjustedRandIndex(w, x)
mclust::adjustedRandIndex(x, y)
mclust::adjustedRandIndex(x, z)
}
\references{
\url{http://faculty.washington.edu/kayee/pca/supp.pdf}
}

