% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_store.R
\name{data_store}
\alias{data_store}
\title{Data Structure for \pkg{hclusttext}}
\usage{
data_store(text, doc.names, min.term.freq = 1, min.doc.len = 1,
  stopwords = tm::stopwords("english"), min.char = 3, max.char = NULL,
  stem = FALSE, denumber = TRUE)
}
\arguments{
\item{text}{A character vector.}

\item{doc.names}{An optional vector of document names corresponding to the
length of \code{text}.}

\item{min.term.freq}{The minimum times a term must appear to be included in
the \code{\link[tm]{DocumentTermMatrix}}.}

\item{min.doc.len}{The minimum words a document must contain to be included
in the data structure (other wise it is stored as a \code{removed} element).}

\item{stopwords}{A vector of stopwords to remove.}

\item{min.char}{The minial length character for retained words.}

\item{max.char}{The maximum length character for retained words.}

\item{stem}{Logical.  If \code{TRUE} the \code{stopwords} will be stemmed.}

\item{denumber}{Logical.  If \code{TRUE} numbers will be excluded.}
}
\value{
Returns a list containing:
\describe{
  \item{dtm}{A tf-idf weighted \code{\link[tm]{DocumentTermMatrix}}}
  \item{text}{The text vector with unanalyzable elements removed}
  \item{removed}{The indices of the removed text elements, i.e., documents not meeting \code{min.doc.len}}
  \item{n.nonsparse}{The length of the non-zero elements}
}
}
\description{
A data structure which stores the text, DocumentTermMatrix, and information
regarding removed text elements which can not be handled by the
\code{hierarchical_cluster} function.  This structure is required because it
documents important meta information, including removed elements, required by
other \pkg{hclustext} functions.  If the user wishes to combine documents
(say by a common grouping variable) it is recomended this be handled by
\code{\link[textshape]{combine}} prior to using \code{data_store}.
}
\examples{
data_store(presidential_debates_2012[["dialogue"]])

## Use `combine` to merge text prior to `data_stare`
library(textshape)
library(dplyr)

dat <- presidential_debates_2012 \%>\%
    dplyr::select(person, time, dialogue) \%>\%
    textshape::combine()

## Elements in `ds` correspond to `dat` grouping vars
(ds <- with(dat, data_store(dialogue)))
dplyr::select(dat, -3)

## Add row names
(ds2 <- with(dat, data_store(dialogue, paste(person, time, sep = "_"))))
rownames(ds2[["dtm"]])

## Get a DocumentTermMatrix
get_dtm(ds2)
}
\keyword{data}
\keyword{structure}

