% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_documents.R
\name{get_documents}
\alias{get_documents}
\alias{get_documents.assign_cluster}
\title{Get Documents Based on Cluster Assignment in \code{assign_cluster}}
\usage{
get_documents(x, ...)

\method{get_documents}{assign_cluster}(x, ...)
}
\arguments{
\item{x}{A \code{\link[hclustext]{assign_cluster}} object.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of \code{\link[base]{vector}}s of document names.
}
\description{
Get the documents associated with each of the k clusters .
}
\examples{
library(dplyr)

mydocuments1 <- presidential_debates_2012 \%>\%
    with(data_store(dialogue, paste(person, time, sep="-"))) \%>\%
    hierarchical_cluster() \%>\%
    assign_cluster(k = 6) \%>\%
    get_documents()

mydocuments1

mydocuments2 <- presidential_debates_2012 \%>\%
    with(data_store(dialogue)) \%>\%
    hierarchical_cluster() \%>\%
    assign_cluster(k = 55) \%>\%
    get_documents()

mydocuments2
}

