% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_removed.R
\name{get_removed}
\alias{get_removed}
\alias{get_removed.data_store}
\alias{get_removed.hierarchical_cluster}
\alias{get_removed.kmeans_cluster}
\alias{get_removed.nmf_cluster}
\alias{get_removed.skmeans_cluster}
\title{Get a Text Stored in a \code{hierarchical_cluster} Object}
\usage{
get_removed(x, ...)

\method{get_removed}{hierarchical_cluster}(x, ...)

\method{get_removed}{kmeans_cluster}(x, ...)

\method{get_removed}{skmeans_cluster}(x, ...)

\method{get_removed}{nmf_cluster}(x, ...)

\method{get_removed}{data_store}(x, ...)
}
\arguments{
\item{x}{A \code{\link[hclustext]{hierarchical_cluster}} object.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector of text strings.
}
\description{
Extract the text supplied to the
\code{\link[hclustext]{hierarchical_cluster}} object.
}
\examples{
library(dplyr)

presidential_debates_2012 \%>\%
    with(data_store(dialogue)) \%>\%
    hierarchical_cluster() \%>\%
    get_removed()
}

