% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_text}
\alias{get_text}
\alias{get_text.assign_cluster}
\alias{get_text.data_store}
\alias{get_text.hierarchical_cluster}
\alias{get_text.kmeans_cluster}
\alias{get_text.nmf_cluster}
\alias{get_text.skmeans_cluster}
\title{Get a Text Stored in Various Objects}
\usage{
get_text(x, ...)

\method{get_text}{hierarchical_cluster}(x, ...)

\method{get_text}{kmeans_cluster}(x, ...)

\method{get_text}{nmf_cluster}(x, ...)

\method{get_text}{skmeans_cluster}(x, ...)

\method{get_text}{data_store}(x, ...)

\method{get_text}{assign_cluster}(x, ...)
}
\arguments{
\item{x}{A \code{\link[hclustext]{hierarchical_cluster}} object.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector or list of text strings.
}
\description{
Extract the text supplied to the
\code{\link[hclustext]{hierarchical_cluster}} object.
}
\examples{
library(dplyr)

presidential_debates_2012 \%>\%
    with(data_store(dialogue)) \%>\%
    hierarchical_cluster() \%>\%
    get_text() \%>\%
    head()
}

