% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_distance.R
\name{jaccard_distance}
\alias{jaccard_distance}
\alias{jaccard_distance.DocumentTermMatrix}
\alias{jaccard_distance.TermDocumentMatrix}
\title{Optimized Computation of Jaccard Distance}
\usage{
jaccard_distance(x, ...)

\method{jaccard_distance}{DocumentTermMatrix}(x, ...)

\method{jaccard_distance}{TermDocumentMatrix}(x, ...)
}
\arguments{
\item{x}{A data type (e.g., \code{\link[tm]{DocumentTermMatrix}} or
\code{\link[tm]{TermDocumentMatrix}}).}

\item{\ldots}{ignored.}
}
\value{
Returns a jaccard distance object of class \code{"dist"}.
}
\description{
Utilizes the \pkg{slam} package to efficiently calculate jaccard distance
on large sparse matrices.
}
\examples{
library(gofastr)
library(dplyr)

out <- presidential_debates_2012 \%>\%
    with(q_dtm(dialogue)) \%>\%
    jaccard_distance()
}
\author{
user41844 of StackOverflow, Dmitriy Selivanov, and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
\url{http://stackoverflow.com/a/36373333/1000343}
\url{http://stats.stackexchange.com/a/89947/7482}
}
\keyword{dissimilarity}
\keyword{jaccard}

