% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_cluster.R
\name{nmf_cluster}
\alias{nmf_cluster}
\alias{nmf_cluster.data_store}
\title{Fit a Non-Negative Matrix Factorization Cluster}
\usage{
nmf_cluster(x, k = k, ...)

\method{nmf_cluster}{data_store}(x, k, ...)
}
\arguments{
\item{x}{A data store object (see \code{\link[clustext]{data_store}}).}

\item{k}{The number of clusters.}

\item{\ldots}{Other arguments passed to \code{\link[rNMF]{rnmf}}.}
}
\value{
Returns an object of class \code{"hclust"}.
}
\description{
Fit a robust non-negative matrix factorization cluster to text data via
\code{\link[rNMF]{rnmf}}.  Prior to distance measures being
calculated the tf-idf (see \code{\link[tm]{weightTfIdf}}) is applied to the
\code{\link[tm]{DocumentTermMatrix}}.
}
\examples{
library(dplyr)

x <- with(
    presidential_debates_2012,
    data_store(dialogue, paste(person, time, sep = "_"))
)


## 6 topic model
model6 <- nmf_cluster(x, k=6)

model6 \%>\%
    assign_cluster()

model6 \%>\%
    assign_cluster() \%>\%
    summary()
\dontrun{
x2 <- presidential_debates_2012 \%>\%
    with(data_store(dialogue))

myfit2 <- nmf_cluster(x2, 55)

assign_cluster(myfit2)

assign_cluster(myfit2) \%>\%
    summary()
}
}

