% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_topic.R
\name{print.as_topic}
\alias{print.as_topic}
\title{Prints an as_topic Object}
\usage{
\method{print}{as_topic}(x, ...)
}
\arguments{
\item{x}{An as_topic object.}

\item{\ldots}{ignored.}
}
\description{
Prints an as_topic object
}

