% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cluster.R
\name{print.assign_cluster}
\alias{print.assign_cluster}
\title{Prints an assign_cluster Object}
\usage{
\method{print}{assign_cluster}(x, ...)
}
\arguments{
\item{x}{An assign_cluster object.}

\item{\ldots}{ignored.}
}
\description{
Prints an assign_cluster object
}

