% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{print.compare}
\alias{print.compare}
\title{Prints a compare Object.}
\usage{
\method{print}{compare}(x, digits = 3, ...)
}
\arguments{
\item{x}{The compare object}

\item{digits}{Number of decimal places to print.}

\item{\ldots}{ignored}
}
\description{
Prints a compare object.
}

