% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_store.R
\name{print.data_store}
\alias{print.data_store}
\title{Prints a data_store Object}
\usage{
\method{print}{data_store}(x, ...)
}
\arguments{
\item{x}{A data_store object.}

\item{\ldots}{ignored.}
}
\description{
Prints a data_store object
}

