% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_documents.R
\name{print.get_documents}
\alias{print.get_documents}
\title{Prints a get_documents Object}
\usage{
\method{print}{get_documents}(x, ...)
}
\arguments{
\item{x}{A get_documents object.}

\item{\ldots}{ignored.}
}
\description{
Prints a get_documents object
}

