% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terms.R
\name{print.get_terms}
\alias{print.get_terms}
\title{Prints a get_terms Object}
\usage{
\method{print}{get_terms}(x, ...)
}
\arguments{
\item{x}{A get_terms object.}

\item{\ldots}{ignored.}
}
\description{
Prints a get_terms object
}

