% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skmeans_cluster.R
\name{skmeans_cluster}
\alias{skmeans_cluster}
\alias{skmeans_cluster.data_store}
\title{Fit a skmean Cluster}
\usage{
skmeans_cluster(x, k, ...)

\method{skmeans_cluster}{data_store}(x, k, ...)
}
\arguments{
\item{x}{A data store object (see \code{\link[clustext]{data_store}}).}

\item{k}{The number of clusters.}

\item{\ldots}{Other arguments passed to \code{\link[skmeans]{skmeans}}.}
}
\value{
Returns an object of class \code{"skmean"}.
}
\description{
Fit a skmean cluster to text data.  Prior to distance measures being
calculated the tf-idf (see \code{\link[tm]{weightTfIdf}}) is applied to the
\code{\link[tm]{DocumentTermMatrix}}.  Cosine dissimilarity is used to generate
the distance matrix supplied to \code{\link[skmeans]{skmeans}}.
}
\examples{
library(dplyr)

x <- with(
    presidential_debates_2012,
    data_store(dialogue, paste(person, time, sep = "_"))
)


## 6 topic model
skmeans_cluster(x, k=6)

skmeans_cluster(x, k=6) \%>\%
    assign_cluster()

skmeans_cluster(x, k=6) \%>\%
    assign_cluster() \%>\%
    summary()

\dontrun{
x2 <- presidential_debates_2012 \%>\%
    with(data_store(dialogue))

myfit2 <- skmeans_cluster(x2, 55)

assign_cluster(myfit2)

assign_cluster(myfit2) \%>\%
    summary()
}
}

