% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cluster_text.R
\name{write_cluster_text}
\alias{read_cluster_text}
\alias{write_cluster_text}
\title{Write/Read Cluster Text for Human Categorization}
\usage{
write_cluster_text(x, path, n.sample = NULL, lead = " * ", ...)

read_cluster_text(path, ...)
}
\arguments{
\item{x}{An \code{assign_cluster} object.}

\item{path}{A pather to the file (.txt) is recommended.}

\item{n.sample}{The length to limit the sample to (default gives all text in the cluster).
Setting this to an integer uses this as the number to randomly sample from.}

\item{lead}{A leading character string prefix to give the cluster text.}

\item{\ldots}{ignored.}
}
\description{
Write cluster text from \code{get_text(assign_cluster(myfit))} to an external 
file for categorization.  After file has been written with 
\code{write_cluster_text} a human coder can assign categories to each cluster.
Simple write the category after the \code{Cluster #:}.  To set a cluster category 
equal to another simply write and equal sign follwed by the other cluster to set
as the same category (e.g., \code{Cluster 10: =5} to set cluster #10 the same as 
cluster #5).  See \code{readLines(system.file("additional/foo_turk.txt", package = "clustext"))} 
for an example.
}
\examples{
library(dplyr)

## Assign Clusters
ca <- presidential_debates_2012 \%>\%
    with(data_store(dialogue)) \%>\%
    hierarchical_cluster() \%>\%
    assign_cluster(k = 7)

## Write Cluster Text for Human Categorization
write_cluster_text(ca)
write_cluster_text(ca, n.sample=10)
write_cluster_text(ca, lead="  -", n.sample=10)

## Read Human Coded Categories Back In
categories_file <- system.file("additional/foo_turk.txt", package = "clustext")
readLines(categories_file)
(categories_key <- read_cluster_text(categories_file))

## Add Categories Back to Original Data Set
categorize(
    data = presidential_debates_2012,
    assign.cluster = ca,
    cluster.key = categories_key
)
}
\seealso{
\code{\link[clustext]{categorize}}
}

