% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Adjusted Rand Index Comaprison Between Algorithms}
\usage{
compare(...)
}
\arguments{
\item{\ldots}{A series of outputs from \code{assign_cluster} for various 
cluster algorithmns.}
}
\value{
Returns a pair-wise comparison matrix of Adjusted Rand Indices for 
algorithm.  Higher Adjusted Rand Index scores indicate higher cluster 
assignment agreement.
}
\description{
An Adjusted Rand Index comparison of the assignments between different
clustering algorithms.
}
\examples{
compare(
    assignments$hierarchical_assignment,
    assignments$kmeans_assignment,
    assignments$nmf_assignment
)
}
\references{
\url{http://faculty.washington.edu/kayee/pca/supp.pdf}
}

