% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_cluster.R
\name{kmeans_cluster}
\alias{kmeans_cluster}
\alias{kmeans_cluster.data_store}
\title{Fit a Kmeans Cluster}
\usage{
kmeans_cluster(x, k = hclustext::approx_k(get_dtm(x)), ...)

\method{kmeans_cluster}{data_store}(x, k = hclustext::approx_k(get_dtm(x)),
  ...)
}
\arguments{
\item{x}{A data type (e.g., \code{\link[tm]{DocumentTermMatrix}} or
\code{\link[tm]{TermDocumentMatrix}}).}

\item{k}{The number of clusters.  Defaults to use \code{approx_k} of the
\code{\link[tm]{DocumentTermMatrix}} produced by \code{data_storage}.}

\item{\ldots}{Other arguments passed to \code{\link[stats]{kmeans}}.}
}
\value{
Returns an object of class \code{"kmeans"}.
}
\description{
Fit a kmeans cluster to text data.  Prior to distance measures being
calculated the tf-idf (see \code{\link[tm]{weightTfIdf}}) is applied to the
\code{\link[tm]{DocumentTermMatrix}}.
}
\examples{
library(dplyr)

x <- with(
    presidential_debates_2012,
    data_store(dialogue, paste(person, time, sep = "_"))
)

## K predicted
kmeans_cluster(x)

## 6 topic model
kmeans_cluster(x, k=6)

kmeans_cluster(x, k=6) \%>\%
    assign_cluster()

kmeans_cluster(x, k=6) \%>\%
    assign_cluster() \%>\%
    summary()

x2 <- presidential_debates_2012 \%>\%
    with(data_store(dialogue))

myfit2 <- kmeans_cluster(x2, 55)

assign_cluster(myfit2)

assign_cluster(myfit2) \%>\%
    summary()
}

