% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_cluster.R
\name{plot.hierarchical_cluster}
\alias{plot.hierarchical_cluster}
\title{Plots a hierarchical_cluster Object}
\usage{
\method{plot}{hierarchical_cluster}(x, k = approx_k(get_dtm(x)), h = NULL,
  color = "red", ...)
}
\arguments{
\item{x}{A hierarchical_cluster object.}

\item{k}{The number of clusters (can supply \code{h} instead).  Defaults to
use \code{approx_k} of the \code{\link[tm]{DocumentTermMatrix}} produced
by \code{data_storage}.  Boxes are drawn around the clusters.}

\item{h}{The height at which to cut the dendrograms (determines number of
clusters).  If this argument is supplied \code{k} is ignored. A line is drawn
showing the cut point on the dendrogram.}

\item{color}{The color to make the cluster boxes (\code{k}) or line (\code{h}).}

\item{\ldots}{Other arguments passed to \code{\link[stats]{rect.hclust}} or
\code{\link[graphics]{abline}}.}
}
\description{
Plots a hierarchical_cluster object
}

